Implement reusability in computing with = sign
Before this commit, the calculator only allowed computing one equation
using the equals sign, requiring user to refresh the page if they want
to compute another one. This commit implements this feature by clearing
the operands and operator after = is pressed.

A `resultMode` variable was also added to track whether the current
value in display is the result of a calculation or user-entered. This
is useful in this particular commit when a user starts to enter a new
equation after they had just computed one -- the display erases the
result of the computed one and starts anew with the digits the user
enters.

Such a flag might be also useful later when adding other calculator
features.