Implement (buggy) logic for running calculation
As of writing, I recognize four cases for when the operator button is
pressed. These were implemented to just get the work flowing!

Known bug: Unable to switch operator in the case of a
running calculation. Example key sequence: 4 x 2 ÷ -
Expected to be able to switch operator (case 4) even when 8 is on the
screen after pressing ÷, but it ends up going into case 3 again to
compute 8÷8.