Rewrite Equals Calculation using OperandB
As part of refactoring the (buggy) solution as of commit c1b7798 to make
more use of operandB, we've rewritten the Equals Calculation first as
this is the most basic case, as opposed to a Running Calculation. The
other parts of solution c1b7798 were deleted or commented out to keep
the testing process clean.