Change implementation of how digits are posted
Once a digit is pressed, the `updatedNumber` is now calculated with
respect to the Current Operand, rather than with respect to the display
contents. This implementation is more consistent with how operands are
updated.

This implementation also makes implementing the Running calculation
feature easier by disemminating the Current Operand's value from the
display constents. And it possibly eliminates the need for a `resultMode`
variable.