Remove the role of `resultMode` in posting a digit
Variable `resultMode` need not be checked anymore when posting digits
due to the new implementation for posting digits - doing so with respect
to the value of `operandA` or `operandB` rather than the value of
display.

`resultMode` is currently only used to check for the case of a Handoff
Calculation when an operator is pressed.