Prevent operator presses from clearing display
Previously, an operator press would clear the display if the user was
not using it to perform a Running Calculation, and not clear the display
if they were. We are considering this an inconsistency. This commit
addresses this by changing operator presses to never clear the display,
only setting or switching the internal `operator` variable when
appropriate.