Remove redundant cast to str when posting digits
This change does not affect the state of the calculator since there
is no feature that relies on the contents of display being a string,
and the contents of display is currently never not just a number in
string form.

Note that handling division by 0 is not yet implemented, so our
treatment of the display's contents might change.