Add gaps between each calculator element
See comments for the reason why a hardcoded value was used for width
instead of delegating calculations to the `flex` property. And variables
were not used since that was not learned yet at this point in the TOP
curriculum.

The height of each button is equal to the width of a `.calc-button`
to make each `.calc-button` square and each `.wide-calc-button`'s height
match with that of the other buttons.

`content-box` was used for the calculator container instead of
`border-box` as that was easier for the width calculation for elements
within the calculator. This does make it harder to arrange elements
outside the calculator, but since this project is the calculator itself,
that difficulty is negligible.