Add keyboard support for calculator
As part of the extra credit.

Note that changes were made to the physical symbols on the calculator
for multiplication and division to match those on a standard QWERTY
keyboard.

Note that using the Enter key as an equivalent to = was withheld from
being implemented due to how focus with buttons works: Once a button
is *clicked*, pressing Enter would activate that button instead of =.