Change user-entered operands to be stored as string
The previous implementation of converting the user-entered operands to
number is largely aesthetic - we do math with numbers after all... But
when planning how to implement entering decimals, we discovered that
this conversion harms us: for instance, a user would not be able to
enter "0.01" as conversion to number at the "0.0" state would revert
the operand to just 0. This commit fixes such cases.

The case of entering 0 first followed by a digit d is accounted for in
this new implementation - operand would still yield "d" instead of "0d".

The math itself is not affected since `operate()` casts both operands
to number before performing math.

It's worth noting that `operandA` is not always string type, as in the
case of a continuous calculation (Handoff Calculation or Running
Calculation).